+++ Infin8.Coapp.Models/Pay_PF_Template.cs
@@ -9,8 +9,8 @@ namespace Infin8.Coapp.Models
         [Key]
         public decimal Pf_Id { get; set; }
         public Nullable<System.DateTime> Wef { get; set; }
-        public int ExGratia_Led_Id { get; set; }
-        public int Bonus_Led_Id { get; set; }
+        public decimal ExGratia_Led_Id { get; set; }
+        public decimal Bonus_Led_Id { get; set; }
         public double PF_Roi { get; set; }
         public bool Pf_Delete { get; set; }
         public decimal Usr_Id { get; set; }
